[20241219] Rename the helper package ## Updates - Renamed a package from `helper` to `helpers`.
