+++ etl-stream/src/main/java/com/github/davidch93/etl/stream/helpers/GcsHelper.java
@@ -1,4 +1,4 @@
-package com.github.davidch93.etl.stream.helper;
+package com.github.davidch93.etl.stream.helpers;
 
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobId;
