+++ etl-stream/src/main/java/com/github/davidch93/etl/stream/pipelines/DataPoolStreamPipeline.java
@@ -4,8 +4,8 @@ import com.github.davidch93.etl.core.pipelines.EtlPipeline;
 import com.github.davidch93.etl.core.schema.Table;
 import com.github.davidch93.etl.stream.config.StreamConfigLoader;
 import com.github.davidch93.etl.stream.config.StreamConfiguration;
-import com.github.davidch93.etl.stream.helper.BigQueryHelper;
-import com.github.davidch93.etl.stream.helper.SchemaHelper;
+import com.github.davidch93.etl.stream.helpers.BigQueryHelper;
+import com.github.davidch93.etl.stream.helpers.SchemaHelper;
 import com.github.davidch93.etl.stream.options.StreamOptions;
 import com.github.davidch93.etl.stream.transformers.DataPoolStream;
 import com.github.davidch93.etl.stream.transformers.Partitioner;
