+++ etl-stream/src/test/java/com/github/davidch93/etl/stream/helpers/GcsHelperTest.java
@@ -1,4 +1,4 @@
-package com.github.davidch93.etl.stream.helper;
+package com.github.davidch93.etl.stream.helpers;
 
 import com.google.cloud.storage.BlobId;
 import com.google.cloud.storage.BlobInfo;
