+++ etl-stream/src/test/java/com/github/davidch93/etl/stream/helpers/SchemaHelperTest.java
@@ -1,4 +1,4 @@
-package com.github.davidch93.etl.stream.helper;
+package com.github.davidch93.etl.stream.helpers;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
